fix input

+++ src/utils.h
@@ -6,7 +6,7 @@
 int initWindow(Chip8* chip8,int textureWidth,int textureHeight);
 void Update(void const* buffer, int pitch);
 void destroyWindow();
-int findIndex(unsigned short arr[], int key);
+int findIndex(unsigned int arr[], int key);
 void handleInput(Chip8* chip8);
 
 #endif
